Implement basic containerized build and arrange for cosmos's secrets management.
 - Implement a buildprocess not using starport CLI so we can take
   advantage of download and binary caches.

 - Implement a container image providing a headless implementation of
   secret service to manage validator keys.

On Linux, cosmos devolves to either the D-Bus secret service API, or
asking for passwords on standard out/in for managing key secrets.  Since
neither running a desktop or implementing stdin/out based interactive
solutions is attractive, we pull in a headless secret-service
implementation, and tie it in with s6-supervise for runtime managmement.
This lets us feed in a master key for the container for secret storage,
and have in implementation that lets us leave the desired secrets
encrypted at rest, in our case, to be provided via a kubernetes secret
projected as a file.